requirements.txt
# ShopPinkki (ros_ws) 통합 Python 의존성
#
# 버전 고정: conda 환경 `jazzy` 기준 (`conda run -n jazzy pip list --format=freeze`)
#
# 설치:
#   pip install -r requirements.txt
#
# ROS 2 바인딩(rclpy, ament_*, 메시지 타입 등)은 pip가 아니라 Ubuntu 패키지와
# colcon 빌드로 설치합니다 (예: sudo apt install ros-jazzy-desktop).
#
# Open-RMF Fleet Adapter는 apt + pip 조합입니다:
#   sudo apt install ros-jazzy-rmf-fleet-adapter ros-jazzy-rmf-traffic ros-jazzy-rmf-task

# --- shoppinkki_core (상태 머신) ---
transitions==0.9.3

# --- control_service (Flask REST + PostgreSQL) ---
Flask==3.1.3
psycopg2-binary==2.9.10

# --- customer_web (Flask + SocketIO) ---
Flask-SocketIO==5.6.1
python-socketio==5.16.1
eventlet==0.41.0
requests==2.33.1

# --- admin_ui (PyQt6; pip가 함께 맞추는 Qt 바인딩) ---
PyQt6==6.11.0
PyQt6-Qt6==6.11.0
PyQt6_sip==13.11.1

# --- 공통 (YAML, RMF HTTP 클라이언트 등) ---
PyYAML==6.0.3

# --- shoppinkki_perception 수치 스택 (jazzy에 설치된 버전) ---
numpy==2.4.4

# --- 개발 / ament 테스트 (jazzy에 설치된 버전) ---
pytest==9.0.2
pytest-repeat==0.9.4

# --- jazzy freeze에 포함된 주요 전이 의존성 (동일 해상도 유지) ---
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
dnspython==2.8.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
python-engineio==4.13.1
simple-websocket==1.1.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
wsproto==1.3.2

# --- conda env jazzy 스냅샷에 없음 (Pi / Docker AI / RMF 전용). 필요 시 별도 설치.
# opencv-python-headless
# ultralytics
# rmf-adapter
# Pillow

# --- Raspberry Pi 전용 (pinky_emotion). x86에서는 보통 생략.
# RPi.GPIO
# spidev

# --- pinky_led / pinky_bringup의 pinkylib 등은 플랫폼 전용 배포물.
